% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.fms_operations.R
\name{delete_notification_channel}
\alias{delete_notification_channel}
\title{Deletes an AWS Firewall Manager association with the IAM role and the Amazon Simple Notification Service (SNS) topic that is used to record AWS Firewall Manager SNS logs}
\usage{
delete_notification_channel()
}
\description{
Deletes an AWS Firewall Manager association with the IAM role and the Amazon Simple Notification Service (SNS) topic that is used to record AWS Firewall Manager SNS logs.
}
\examples{

}
