% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.fms_operations.R
\name{delete_policy}
\alias{delete_policy}
\title{Permanently deletes an AWS Firewall Manager policy}
\usage{
delete_policy(PolicyId, DeleteAllPolicyResources = NULL)
}
\arguments{
\item{PolicyId}{[required] The ID of the policy that you want to delete. \code{PolicyId} is returned by \code{PutPolicy} and by \code{ListPolicies}.}

\item{DeleteAllPolicyResources}{If \code{True}, the request will also delete all web ACLs in this policy. Associated resources will no longer be protected by web ACLs in this policy.}
}
\description{
Permanently deletes an AWS Firewall Manager policy.
}
\section{Accepted Parameters}{
\preformatted{delete_policy(
  PolicyId = "string",
  DeleteAllPolicyResources = TRUE|FALSE
)
}
}

