% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.fms_operations.R
\name{disassociate_admin_account}
\alias{disassociate_admin_account}
\title{Disassociates the account that has been set as the AWS Firewall Manager administrator account}
\usage{
disassociate_admin_account()
}
\description{
Disassociates the account that has been set as the AWS Firewall Manager administrator account. To set a different account as the administrator account, you must submit an \code{AssociateAdminAccount} request .
}
\section{Accepted Parameters}{
\preformatted{disassociate_admin_account()
}
}

