% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.fms_operations.R
\name{get_admin_account}
\alias{get_admin_account}
\title{Returns the AWS Organizations master account that is associated with AWS Firewall Manager as the AWS Firewall Manager administrator}
\usage{
get_admin_account()
}
\description{
Returns the AWS Organizations master account that is associated with AWS Firewall Manager as the AWS Firewall Manager administrator.
}
\section{Accepted Parameters}{
\preformatted{get_admin_account()
}
}

