% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.fms_operations.R
\name{get_notification_channel}
\alias{get_notification_channel}
\title{Returns information about the Amazon Simple Notification Service (SNS) topic that is used to record AWS Firewall Manager SNS logs}
\usage{
get_notification_channel()
}
\description{
Returns information about the Amazon Simple Notification Service (SNS) topic that is used to record AWS Firewall Manager SNS logs.
}
\section{Accepted Parameters}{
\preformatted{get_notification_channel()
}
}

