% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.fms_operations.R
\name{get_policy}
\alias{get_policy}
\title{Returns information about the specified AWS Firewall Manager policy}
\usage{
get_policy(PolicyId)
}
\arguments{
\item{PolicyId}{The ID of the AWS Firewall Manager policy that you want the details for.}
}
\description{
Returns information about the specified AWS Firewall Manager policy.
}
\examples{

}
