% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.fms_operations.R
\name{delete_policy}
\alias{delete_policy}
\title{Permanently deletes an AWS Firewall Manager policy}
\usage{
delete_policy(PolicyId)
}
\arguments{
\item{PolicyId}{[required] The ID of the policy that you want to delete. \code{PolicyId} is returned by \code{PutPolicy} and by \code{ListPolicies}.}
}
\description{
Permanently deletes an AWS Firewall Manager policy.
}
\section{Accepted Parameters}{
\preformatted{delete_policy(
  PolicyId = "string"
)
}
}

