% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.fms_operations.R
\name{disassociate_admin_account}
\alias{disassociate_admin_account}
\title{Disassociates the account that has been set as the AWS Firewall Manager administrator account}
\usage{
disassociate_admin_account()
}
\description{
Disassociates the account that has been set as the AWS Firewall Manager administrator account. You will need to submit an \code{AssociateAdminAccount} request to set a new account as the AWS Firewall administrator.
}
\examples{

}
