% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.fms_operations.R
\name{put_notification_channel}
\alias{put_notification_channel}
\title{Designates the IAM role and Amazon Simple Notification Service (SNS) topic that AWS Firewall Manager uses to record SNS logs}
\usage{
put_notification_channel(SnsTopicArn, SnsRoleName)
}
\arguments{
\item{SnsTopicArn}{The Amazon Resource Name (ARN) of the SNS topic that collects notifications from AWS Firewall Manager.}

\item{SnsRoleName}{The Amazon Resource Name (ARN) of the IAM role that allows Amazon SNS to record AWS Firewall Manager activity.}
}
\description{
Designates the IAM role and Amazon Simple Notification Service (SNS) topic that AWS Firewall Manager uses to record SNS logs.
}
\examples{

}
