% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.fms_operations.R
\name{put_policy}
\alias{put_policy}
\title{Creates an AWS Firewall Manager policy}
\usage{
put_policy(Policy)
}
\arguments{
\item{Policy}{The details of the AWS Firewall Manager policy to be created.}
}
\description{
Creates an AWS Firewall Manager policy.
}
\examples{

}
